@@ -4,7 +4,7 @@ version = "0.1.0-pre.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 license = "MIT"
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 readme = "README.md"
 repository = "https://github.com/artichoke/artichoke"
 documentation = "https://artichoke.github.io/artichoke/artichoke/"
