@@ -3,7 +3,7 @@ name = "artichoke-backend"
 version = "0.6.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 description = "Embeddable VM implementation for Artichoke Ruby"
 repository = "https://github.com/artichoke/artichoke"
 readme = "README.md"
