@@ -3,7 +3,7 @@ name = "artichoke-core"
 version = "0.10.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 description = "Core traits for implementing an Artichoke Ruby interpreter"
 repository = "https://github.com/artichoke/artichoke"
 readme = "README.md"
