@@ -3,7 +3,7 @@ name = "artichoke-load-path"
 version = "0.1.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 description = "Source and extension loaders for a managing a Ruby $LOAD_PATH"
 repository = "https://github.com/artichoke/artichoke"
 readme = "README.md"
