@@ -3,7 +3,7 @@ name = "scolapasta-hex"
 version = "0.1.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 description = """
 no_std hexadecimal encoding utility package for Artichoke Ruby.
 """
