@@ -3,7 +3,7 @@ name = "spec-runner"
 version = "0.6.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 description = "Binary for running Ruby Specs with Artichoke Ruby"
 repository = "https://github.com/artichoke/artichoke"
 readme = "README.md"
