@@ -3,7 +3,7 @@ name = "spinoso-array"
 version = "0.7.0"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 edition = "2021"
-rust-version = "1.57.0"
+rust-version = "1.58.0"
 description = """
 Growable vector backends for the Ruby Array core type in Artichoke Ruby
 """
