@@ -363,12 +363,12 @@ mod tests {
     #[test]
     fn fmt_debug_does_not_leak_seed() {
         let random = Random::with_seed(874);
-        let debug = format!("{:?}", random);
+        let debug = format!("{random:?}");
         assert!(!debug.contains("894"));
         assert_eq!(debug, "Random {}");
 
         let random = Random::with_seed(123_456);
-        let debug = format!("{:?}", random);
+        let debug = format!("{random:?}");
         assert!(!debug.contains("123456"));
         assert_eq!(debug, "Random {}");
     }
